internal/abi: avoid rebuilding all of std in TestFuncPCCompileError
Use 'go list -export' directly instead of creating an importcfg for
the entire standard library.

After #47257, a builder running a test does not necessarily have the
entire standard library already cached, especially when running tests
in sharded mode. testenv.WriteImportcfg writes an importcfg for the
entire standard library, but this test only needs a tiny subset —
"internal/abi" and its transitive dependencies. We can supply that
much more efficiently.

This reduces the time to run the test with a cold build cache on my
workstation from ~16s to ~0.05s.

Fixes #58248.

Change-Id: I9be6b61ae6e28b75b53af85207c281bb93b9346f