cmd/compile: rename EvalConst to EvalExpr
Follow up discussion in CL 465096, all constants evaluation is already
done by types2, downstream parts of the compiler should not have any
constant evaluation. EvalConst has the ability to handle most of the
constant expression in non-constant semantic already. The only missing
part is the handling of OCONV, which requires truncating or sign
extending for intenger conversion. This can be done by copying the
"truncate" function added in CL 451376. For the current cases that
report an error or fatal, just return the original expression and let
the runtime error trigger.

While at it, rename EvalConst to EvalExpr to distinguish with the
constants evaluation happend in types2.

Updates #58293
Updates #58339
Fixes #58439

Change-Id: I74068d99c245938e576afe9460cbd2b39677bbff