runtime: disable huge pages for persistentalloc chunks by default
This change modifies persistentalloc internals to not back chunks with
huge pages to begin with. It only backs chunks with huge pages once the
heap reaches a certain size, like other heap metadata.

The reason we enable huge pages based on heap size for persistentalloc
is because, although anything could go into persistentalloc, a lot of
the memory going in there comes from metadata that scales with peak heap
footprint. For example, heapArena structures are peristentalloc'd and
they each represent some region of address space mapped for the heap.
Another example is fixalloc chunks for mspans, the number of which tends
to grow proportionally to heap size (less true in general, but not
totally off-base). Same for spanSet blocks.

Notably, however, this CL alone isn't enough to capture heapArenas
because they exceed the maximum allocation size for persistentalloc, so
they're mapped directly. A follow-up CL will handle this case.

Change-Id: Idef1510ee4a19990d013f2f0d6ea746332f38d61