runtime: track persistentalloc mappings for large allocations
Currently large persistentalloc'd values (such as heapArena) aren't
tracked by persistentAlloc at all. Technically this means that
inPersistentAlloc has an incomplete implementation, but that likely
doesn't matter very much. What makes a bigger difference is that these
mappings don't participate in the huge page policy the previous CL set
up.

Fix this by tracking persistentalloc mappings in the same linked list as
mappings for small objects. This is a little tricky to do because these
mappings for large allocations rely on being able to provide a pretty
high alignment for persistent allocations, so we can't exactly place a
node link in the first word like for small chunks (not without wasting a
lot of memory).

That leaves the option of placing the node at the end of the mapping.
This works fine, but comes with the caveat that now nice
power-of-two-sized persistent allocations whose mappings come from the
operating system will now waste a whole physical page for this node. In
practice, however, this is not a problem, since such power-of-two sized
persistent allocations don't actually ever happen, at least not when
running all.bash, so there's always a little bit of space we can use.

Change-Id: Ife4934981ec2ece4d32cdc61b4d36a00bab026f5