cmd/compile: add new SSA pass which rewrites recursive tail calls into loops
Unlike previous patches and discussions, this patch dodges the
runtime.{Stack,Caller{,s}} arguments by conservatively applying the
optimization in a way that any attempts at observing a change in behaviour
(such as by noticing missing frames in runtime.Callers) will turn off the
optimization in this function.

This is hit 49 times when building toolchain3, tools and std.

On each occurance this saves a call and may unlock further optimisations.

The build time cost should be fairly minimal as long as the optimization is
not triggered because it is designed to fail fast.

Change-Id: I0087ec42ee3f8c79febb0f23bfe9f4193a9ef0c8