strconv: add Atoi test cases with characters less than '0'
Atoi validates input is digits with ch -= '0' and ch > 9. Because byte is an
unsigned value, if ch is less than '0', the subtraction causes it to become
a large number greater than 9. This is a neat trick to avoid comparing to
two values, '0' and '9' to validate but none of the test cases seem to
include a character less than '0'. While chance of regression seems slim,
checking the behavior for such an overflowing character can prevent it
if something were to change.

Change-Id: I35c0fe4e7e11b694ef07687d5e0532a2ab9ba213
GitHub-Last-Rev: e0a28f42367c2685d80bcc5c45294c0d2eae77fe
GitHub-Pull-Request: golang/go#58575