cmd/compile, cmd/link, runtime, internal: 16-byte-align slices experiment
DO NOT SUBMIT

This overloads swaplencap to also imply wider alignment,
that might be a mistake, but this is not something I want
to mix+match.

Does not work yet because of a problem passing parameters
to ABI0 functions; a slice parameter takes an extra word in
memory.  Still working on the best approach to this.

Change-Id: I7bb475fa77c853d414d137599092df8efe80589b