go/types, types2: use new type inference algorithm exclusively
The primary change is that type inference now always reports
an error if a unification step fails (rather than ignoring that
case, see infer2.go). This brings the implementation closer to
the description in #58650; but the implementation is more direct
by always maintaining a simple (type parameter => type) mapping.

To make this work, there are two small but subtle changes in the
unifier:

1) When deciding whether to proceed with the underlying type of
   a defined type, we also use the underlying type if the other
   type is a basic type (switch from !hasName(x) to isTypeLit(x)
   in unifier.go). This makes the case in issue #53650 work out.

2) When we unify against an unbound type parameter, we always
   proceed with its core type because that's the only thing we
   can do.

The remaining changes are comment and test adjustments. Because
the new logic now results in failing type inference where before
we ended up with incorrectly inferred types that then lead to
instantiation failure, a handful of error messages changed. But
we still have the same number of errors for the same programs.

Also, because type inference now produces different results, we
cannot compare easily against infer1 anymore (also infer1 won't
work correctly anymore due to the changes in the unifier). This
comparison (together with infer1) is now disabled.

Because some errors and their positions have changed, we need a
slightly larger error position tolerance for types2 (which produces
less accurate error positions than go/types), hence the change in
types2/check_test.go.

Change-Id: Id796eea42f1b706a248843ad855d9d429d077bd1