cryptobyte: disambiguate empty fields from errors
The `read()` function underpins many of the other functions within `cryptobyte`.  However it returned only `[]byte`, which in come cases (i.e. a zero length field) could be empty despite there being no error (such as running out of bytes to process).  Many functions (e.g. `ReadUint16LengthPrefixed`) only perform a check for `nil` or similar to determine a success/fail state, and so would return failure (`bool == false`) in the event of cases such as a zero length field, when there is still more data to follow.

When processing something such as a TLS handshake, for example, it is not uncommon to encounter a 0 length field, with valid data following.  However when using `cryptobyte` to check this the `bool` will be set to false, and there is no other form of error checking available.

Therefore the supplied patch modifies the `read()` function to return `([]byte, error)` instead, so that an empty `[]byte` can be returned along with a success status for such cases.  The rest of the suggested change consists of modifications to the rest of the package to account for this change.

These changes are self contained within the package and so should merely fix the bug without changes required elsewhere.

Change-Id: Idc646e5276b33aefa0a7b31706f4d4d438c26620
GitHub-Last-Rev: 6116cad25845bab64c7bf5043613e05fe324d7b9
GitHub-Pull-Request: golang/go#58677