net: mptcp: implement dialMPTCP()
This function is called when the user has requested MPTCP via
SetMultipathTCP() in the Dialer.

This new function fallsback to dialTCP() on operating systems that do
not support MPTCP or if MPTCP is not supported.

Note that this new dialMPTCP() function returns a TCPConn object, like
dialTCP(). A new MPTCPConn object using the following composition could
have been returned:

    type MPTCPConn struct {
        *TCPConn
    }

But the drawback is that if MPTCP is used by default one day (see #56539
issue on GitHub), Dial() will return a different object: this new
MPTCPConn type instead of the previously expected TCPConn. This can
cause issues for apps checking the returned object.

Updates #56539

Link: https://github.com/golang/go/issues/56539
Signed-off-by: Gregory Detal <gregory.detal@tessares.net>
Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>
Change-Id: I0f9b5b81f630b39142bdd553d4f1b4c775f1dff0