net: mptcp: add MPTCP Sock
This currently defines an internal function supportsMultipathTCP() which
reports whether MPTCP[1] is supported on the current platform.

Only Linux is supported here. The check on Linux is performed by
attemting to create an MPTCP socket and look at the returned error. If
the protocol is not supported, EINVAL (kernel < 5.6) or EPROTONOSUPPORT
(kernel >= 5.6) is returned and there is no point to try again. This
test is only done once.

[1] https://www.rfc-editor.org/rfc/rfc8684.html

Updates #56539

Link: https://github.com/golang/go/issues/56539
Signed-off-by: Gregory Detal <gregory.detal@tessares.net>
Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>
Change-Id: Ic84fe85aad887a2be4556a898e649bf6b6f12f03