net: mptcp: implement listenMPTCP()
Similar to dialMPTCP(), this function is called when the user has
requested MPTCP via SetMultipathTCP() in the ListenConfig.

This function fallsback to listenTCP() on operating systems that do not
support MPTCP or if MPTCP is not supported.

Note that this new listenMPTCP() function returns a TCPListener object
and not a new MPTCP dedicated one. The reasons are similar as explained
in the parent commit introducing dialTCP(): if MPTCP is used by default
later, Listen() will return a different object that could break existing
applications expecting TCPListener.

Updates #56539

Link: https://github.com/golang/go/issues/56539
Signed-off-by: Gregory Detal <gregory.detal@tessares.net>
Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>
Change-Id: I010f1d87f921bbac9e157cef2212c51917852353