net: mptcp: implement listenMPTCP
Similar to dialMPTCP(), this function is called when the user has
requested MPTCP via UseMultipathTCP() in the Listener.

This function fallsback to listenTCP() on operating systems that do not
support MPTCP or if MPTCP is not supported.

Note that this new function returns a TCPListener object and not a new
MPTCP dedicated one. There are more explanations about that in the
parent commit introducing dialTCP().

Updates #56539

Link: https://github.com/golang/go/issues/56539
Signed-off-by: Gregory Detal <gregory.detal@tessares.net>
Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>
Change-Id: I010f1d87f921bbac9e157cef2212c51917852353