encoding/json: improve Marshal memory utilization
Previously, Marshal (and Encoder.Encode) used a pooled encodeState
that included an embedded bytes.Buffer.
This is an incorrect use of sync.Pool (see #23199).
This results in problems when a high frequency stream of tiny Marshals
is interspersed with a low frequency stream of huge Marshals.
The occasional huge Marshal will cause bytes.Buffer to be grown in size
to accomodate the large output, but never be reduced in size.
However, the continuous stream of tiny Marshals will keep the buffers alive,
preventing the GC from being able to reclaim the under-utilized capacity.

Previous attempts to trivially cap the buffer capacity (see CL 136035)
resulted in unacceptably poor performance in marshaling large outputs.

We fix this problem by switching from bytes.Buffer to a custom pooledBuffer
implementation that uses a series of pooled buffer segments.
The content of the buffer is the concatenation of all the buffer segments.
Using buffer segments allows the implementation to more readily return
unused buffers to the pool, and allow the GC to reclaim them if necessary.

The semantics of the Marshal function requires that the returned buffer
be a unique copy owned by the caller.
With the segmented buffer, we use bytes.Join to join all the segments,
while the previous code did the semantic equivalent of bytes.Clone.

Performance:

	name                       old time/op      new time/op      delta
	LargeOutput                  1.85s ± 3%       1.31s ± 2%     -29.25%  (p=0.000 n=10+8)
	CodeEncoder                  394µs ± 1%       382µs ± 2%      -2.95%  (p=0.000 n=9+10)
	CodeEncoderError             437µs ± 1%       428µs ± 1%      -1.97%  (p=0.000 n=9+9)
	AllocationWastage            396µs ± 2%       476µs ± 1%     +20.16%  (p=0.000 n=10+8)
	CodeMarshal                  534µs ± 1%       712µs ± 3%     +33.31%  (p=0.000 n=10+10)
	CodeMarshalError             694µs ± 2%       903µs ± 4%     +29.98%  (p=0.000 n=10+10)
	MarshalBytes/32              170ns ± 3%       164ns ± 1%      -3.22%  (p=0.000 n=10+10)
	MarshalBytes/256             387ns ± 2%       381ns ± 1%      -1.63%  (p=0.000 n=10+10)
	MarshalBytes/4096           3.93µs ± 2%      3.91µs ± 1%        ~     (p=0.184 n=10+10)
	MarshalBytesError/32         372µs ± 1%       370µs ± 0%      -0.71%  (p=0.002 n=10+8)
	MarshalBytesError/256        372µs ± 1%       370µs ± 1%        ~     (p=0.133 n=9+10)
	MarshalBytesError/4096       379µs ± 1%       375µs ± 1%      -1.03%  (p=0.007 n=10+10)
	EncodeMarshaler             22.8ns ± 2%      24.4ns ± 2%      +7.06%  (p=0.000 n=8+9)
	EncoderEncode               12.5ns ± 3%      13.6ns ± 8%      +8.23%  (p=0.005 n=8+8)

	name                       old alloc/op     new alloc/op     delta
	LargeOutput                 1.55GB ± 0%      0.97GB ± 0%     -37.61%  (p=0.000 n=10+7)
	CodeEncoder                  4.00B ± 0%    3608.40B ±77%  +90110.00%  (p=0.000 n=9+10)
	CodeEncoderError              141B ± 2%      2605B ±207%   +1751.22%  (p=0.000 n=10+9)
	AllocationWastage            2.87B ±30%    2017.10B ±54%  +70060.00%  (p=0.000 n=8+10)
	CodeMarshal                 1.96MB ± 2%      1.99MB ± 1%      +1.52%  (p=0.000 n=10+10)
	CodeMarshalError            2.04MB ± 2%      2.00MB ± 1%      -1.89%  (p=0.000 n=10+9)
	EncodeMarshaler              4.00B ± 0%       4.00B ± 0%        ~     (all equal)
	EncoderEncode                0.00B            0.00B             ~     (all equal)

	name                       old allocs/op    new allocs/op    delta
	LargeOutput                   44.8 ±23%     19649.6 ± 0%  +43760.71%  (p=0.000 n=10+10)
	CodeEncoder                   0.00            47.00 ± 0%       +Inf%  (p=0.000 n=10+10)
	CodeEncoderError              4.00 ± 0%       51.00 ± 0%   +1175.00%  (p=0.000 n=10+10)
	AllocationWastage             0.00            52.00 ± 0%       +Inf%  (p=0.000 n=10+10)
	CodeMarshal                   1.00 ± 0%       53.00 ± 0%   +5200.00%  (p=0.000 n=10+8)
	CodeMarshalError              6.30 ±11%       59.40 ± 1%    +842.86%  (p=0.000 n=10+10)
	EncodeMarshaler               1.00 ± 0%        1.00 ± 0%        ~     (all equal)
	EncoderEncode                 0.00             0.00             ~     (all equal)

	name                       old wasted_b/op  new wasted_b/op  delta
	AllocationWastage            2.25M ± 0%       0.16M ± 0%     -92.88%  (p=0.000 n=10+10)

There is no notable performance difference on small output (i.e., < 64KiB).
For larger outputs (e.g., CodeMarshal), we do see a performance degradation.
This is unavoidable as having a single contiguous buffer that is large enough
is always better for CPU performance for many reasons.
However, the nature of the Marshal call makes it such that it is impossible
to know what size buffer we need prior to marshaling the value itself.
The loss of performance is deemed acceptable in order to gain
stronger guarantees on the maximum amount of heap used.

Part of the cost of using bytes.Join over bytes.Clone is that
bytes.Clone can rely on compiler optimizations that avoid unnecessarily
zeroing the buffer for the portion that is gauranteed to be overwritten.
The bytes.Join function does not benefit from that compiler optimization.
In CL 456336, bytes.Join is optimized to also avoid the zeroing.
This results in the performance loss being reduced to:

	name                       old time/op      new time/op      delta
	CodeMarshal                  534µs ± 1%       633µs ± 4%     +18.46%  (p=0.000 n=10+10)
	CodeMarshalError             694µs ± 2%       789µs ± 3%     +13.64%  (p=0.000 n=10+9)

This is a reduction of performance loss from +33% down to +18%,
but is still much better than the performance loss of +72%
without the use of any pooled buffers at all.

The idea of a segmented buffer came about during discussion with
Daniel Smith on CL 455236 as a counter proposal
to using pooled contiguous buffers.
The benchmarks come from Daniel Smith's alternate implementation in CL 455776.
Huge thanks goes to Daniel Smith who did a bulk majority of the work
to prove that segmented buffers was a viable approach.
The implementation in this CL is a simplified version of CL 455776.

Fixes #27735

Change-Id: Ib23f00eb64615b21b1ddddf9fadc6498bcefa935