fix/fuzz: output int32(v)/rune(q) in failing log
fix/fuzz: output int32(v)/rune(q) in failing log

When fuzz testing outputs failing input into testdata/<log>, it needs to
marshal the input, In go, int32 and rune are the same datatype, so it
cannot differentiate the two types in typeswitch.

Before, when the data is a valid unicode point, it will outputs `rune(%q)`
in the log, but the real datatype maybe a int32. In this case, we have
to query the ASCII or unicode table to get the int32 value. It's
inconvenient.

So, here we output `int32(%v)/rune(%q)` to solve this.

close #58824

Change-Id: Ia36f517770203faccdd9a8a2f5fc715213f1f1d5
GitHub-Last-Rev: a38e4ec203465fadfe07626f71b88d81cf653d8c
GitHub-Pull-Request: golang/go#58827