runtime: Protect against (possible) nil-pointer dereference
In cases where user applications add items to the list of overlapped IO
operations that may be collected in a GetQueuedCompletionStatusEx, it
may not always be possible/correct to assume that the
`windows.Overlapped` `struct`ure is a `net_poll`. In those cases, we
have to be careful.

For golang.org/go#58870

(Mutual) Depends on https://go.dev/cl/473163

Change-Id: I3e0aa1ece9d50bedfeced29846bb8edbb48c8c8d