cmd/go: restore 'extend path shortening to all paths in messages'
The problem requiring a revert was a problem in the test harness
running on Windows, and would not have worked if there were a test
for the existing shortening behavior.  I hand-verified that the code
works properly on Windows when run normally.

original was CL 465805
revert was CL 473795

this is the original, plus a testscript skip for the extra path
shortening on Windows.

Change-Id: I786e875e704c2d7018c8248960f2ff7188cac3ef