cmd/link: do not generate label symbols for darwin/arm64
In the linker, in external linking mode, we generate label symbols
for relocatation targets that cannot be represented in the
relocation format of the host object (e.g. for large addend, or
even for just having an addend). On darwin/arm64, CALL relocations
can actuall take a (not too big) addend, so there is no need to
generate a label symbol for it.

Change-Id: I38aed6b62a0496c277c589b5accbbef6aace8dd5