runtime: use frame pointer unwinding for trace events
Change tracer to use frame pointer unwinding by default on amd64. The
expansion of inline frames is delayed until the stack table is dumped at
the end of the trace. This requires storing the skip argument in the
stack table, which now resides in pcBuf[0]. For stacks that are not
produced by traceStackID (e.g. CPU samples), a logicalStackSentinel
value in pcBuf[0] indicates that no inline expansion is needed.

Add new GODEBUG=tracefpunwindoff=1 option to use the old unwinder if
needed.

The motivation for this is to significantly reduce the CPU overhead of
the tracer.

goos: linux
goarch: amd64
pkg: runtime
cpu: Intel(R) Xeon(R) Platinum 8375C CPU @ 2.90GHz
               │ fpunwindoff1.txt │       fpunwindoff0.txt       │
               │      sec/op      │   sec/op     vs base         │
PingPongHog-32       3596.5n ± 0%   806.5n ± 0%  -77.58% (n=100)

For #16638

Change-Id: I2928a2fcd8779a31c45ce0f2fbcc0179641190bb