runtime: crash with RaiseFailFastException on Windows
DO NOT REVIEW.
TODO: investigate if this makes sense even if WER is not used.

This CL instructs Windows to crash using RaiseFailFastException [1]
instead of using exit(2) when GOTRACEBACK=crash is set. This way the
shell exist status will contain the expected code instead of a
hardcoded 2, we remove a long-standing TODO.

I've seen this improvement opportunity while investigating how #57441
(WER support) could be implemented. While it is true that using
RaiseFailFastException is necessary to implement WER, this CL does not
change the fact that the Go runtime is unconditionally disabling WER
via SetErrorMode, so it does not change the WER behavior.

[1] https://learn.microsoft.com/en-us/windows/win32/api/errhandlingapi/nf-errhandlingapi-raisefailfastexception

Change-Id: I45669d619fbbd2f6413ce5e5f08425ed1d9aeb64